fix bug after self test

+++ KLTN.Core/ProductServices/DTOs/ProductOnSaleDTO.cs
@@ -8,7 +8,7 @@
         public long ProductNftId { get; set; }
         public string ProductHahIPFS { get; set; }
         public long AmountOnSale { get; set; }
-        public long PriceOfOneItem { get; set; }
+        public decimal PriceOfOneItem { get; set; }
         public string ProductTypeName { get; set; }
         public string ProductDescription { get; set; }
         public string SaleAddress { get; set; }
