+++ KLTN.Core/ProductServices/DTOs/ProductStudentListOnSaleDTO.cs
@@ -4,7 +4,7 @@
     {
         public long ProductNftId { get; set; }
         public long AmountOnSale { get; set; }
-        public long PriceOfOneItem { get; set; }
+        public decimal PriceOfOneItem { get; set; }
         public string SaleAddress { get; set; }
     }
 }
